+++ lib/active_record_compose/transaction_support.rb
@@ -6,6 +6,14 @@ module ActiveRecordCompose
     extend ActiveSupport::Concern
     include ActiveRecord::Transactions
 
+    included do
+      # ActiveRecord::Transactions is defined so that methods such as save,
+      # destroy and touch are wrapped with_transaction_returning_status.
+      # However, ActiveRecordCompose::Model does not support destroy and touch, and
+      # we want to keep these operations as undefined behavior, so we remove the definition here.
+      undef_method :destroy, :touch
+    end
+
     module ClassMethods
       def lease_connection
         if ar_class.respond_to?(:lease_connection)